Reproducible paper explaining kirchoff modeling #13
This SConstruct generates a reproducible paper automatically after
running the numerical experiment of kirchoff modeling available in
the ./scons directory.

It uses references available in mybib.bib file and paper.tex is the
Latex file template.

More information about how to generate reproducicle papers and dependencies
are available in the Madagascar website:

http://www.ahay.org/wiki/Reproducible_computational_experiments_using_SCons