Semblance and nonHyperbolicApp functions tests #18
Run this test with 'make semblance'. This test requires the
test data cube generated in test_data_cube directory.

The Makefile calls generate_data_cube taget to generate the
datacube with scons in the test_data_cube. After that it runs
the unit tests in the Test_semblance.c file using Unity framework.